+++ cycles-ledger/src/main.rs
@@ -54,14 +54,6 @@ fn post_upgrade(ledger_args: Option<LedgerArgs>) {
                     config.index_id = change_index_id.into();
                 })
             }
-            if let Some(change_logo) = upgrade_args.logo {
-                mutate_config(|config| {
-                    match change_logo {
-                        endpoints::ChangeLogo::Unset => config.logo = None,
-                        endpoints::ChangeLogo::SetTo(logo) => config.logo = Some(logo),
-                    }
-                })
-            }
         }
         None | Some(LedgerArgs::Upgrade(None)) => {},
         _ =>
@@ -169,12 +161,10 @@ fn icrc1_metadata() -> Vec<(String, MetadataValue)> {
             "dfn:max_blocks_per_request",
             read_config(|config| config.max_blocks_per_request),
         ),
+        MV::entry("icrc1:logo", config::TOKEN_LOGO.as_bytes()),
     ];
     if let Some(index_id) = read_config(|config| config.index_id) {
         metadata.push(MV::entry("dfn:index_id", index_id.as_slice()))
-    }
-    if let Some(logo) = read_config(|config| config.logo.clone()) {
-        metadata.push(MV::entry("icrc1:logo", logo))
     }
     metadata
 }