+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 # CHANGELOG
 
 ## [Unreleased] - ReleaseDate
-* Allow configuring a logo in init args or upgrade args.
+* Added the logo to the metadata value `icrc1:logo`.
 
 ## [0.2.1] - 2023-09-20